Extract expression detection methods to shared module
- Create ExpressionUtils module for ruby_expression? and bash_command? methods
- Include module in both ConditionalStep and BaseIterationStep
- Remove duplicate method definitions
- Add extract_expression and extract_command utility methods

This completes the refactoring that was mentioned in the initial
summary but had been missed.
